refactor the codes and made some chnages to the design

+++ pages/api/products/[id].js
@@ -7,8 +7,11 @@ export default async function handler(req, res) {
   //extracting query:{id} that is coming from the [id].js in the product folder when when fetching data from the database
   const {
     method,
-    query: { id }
+    query: { id },
+    cookies
   } = req;
+
+  const token = cookies.token;
 
   dbConnect();
 
@@ -23,6 +26,10 @@ export default async function handler(req, res) {
   }
 
   if (method === "PUT") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authorized!");
+    }
+
     try {
       const product = await Product.findByIdAndUpdate(id, req.body, {
         new: true
@@ -34,6 +41,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.token) {
+      return res.status(401).json("Not authorized!");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("The product has been deleted.");
